// frontend/src/client/models/AddManualPageResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Response schema for manually adding a page.
 */
export type AddManualPageResponse = {
    page_id: string;
    url: string;
    slug: string;
    status: string;
    message: string;
};